+++ threedi_schematisation_editor/__init__.py
@@ -3,6 +3,7 @@ import os.path
 import platform
 from collections import defaultdict
 from pathlib import Path
+from typing import Optional
 
 import pyplugin_installer
 from qgis.core import (
@@ -347,7 +348,11 @@ class ThreediSchematisationEditorPlugin:
                 {self.THREEDI_GPKG_VAR_NAMES: project_model_gpkgs_str}
             )
 
-    def load_schematisation(self, model_gpkg=None):
+    def load_schematisation(self, model_gpkg=None, parents: Optional[list[str]] = None):
+        """
+        :param model_gpkg: path to schematisation; if omitted the user can select via the UI
+        :param parents: parent groups that are used to create the root for the schematisation in the layers view
+        """
         if not model_gpkg:
             schematisation_loader = LoadSchematisationDialog(self.uc)
             result = schematisation_loader.exec_()
@@ -389,7 +394,7 @@ class ThreediSchematisationEditorPlugin:
                 self.uc.show_warn(warn_msg, None, "Rana Schematisation Editor")
                 return
 
-        lm = LayersManager(self.iface, self.uc, model_gpkg)
+        lm = LayersManager(self.iface, self.uc, model_gpkg, parents=parents)
         if lm in self.workspace_context_manager:
             self.uc.clear_message_bar()
             warn_msg = "Selected schematisation is already loaded. Loading canceled."